CLOUD-55042 Flow simplification, collect actions in one file
