+++ core/src/main/java/com/sequenceiq/cloudbreak/core/flow2/FlowState.java
@@ -2,6 +2,7 @@ package com.sequenceiq.cloudbreak.core.flow2;
 
 public interface FlowState<S extends FlowState, E extends FlowEvent> {
     Class<?> action();
+    String name();
     S failureState();
     E failureEvent();
 }
